% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_paragraphs_to_doc.R
\name{add_paragraphs_to_doc}
\alias{add_paragraphs_to_doc}
\title{Add paragraphs to a Word document object}
\usage{
add_paragraphs_to_doc(doc, text)
}
\arguments{
\item{doc}{Word document object to be modified}

\item{text}{Text to be split and inserted into the Word document object}
}
\value{
A modified Word document object with the inserted paragraphs
}
\description{
This function takes a Word document object and a text string as input. It splits the text into paragraphs based on newline characters and inserts them into the Word document object.
}
\examples{
\dontrun{
library(officer)

my_doc <- read_docx() # Create a new Word document
text_to_insert <- "This is the first paragraph.\nThis is the second paragraph."
updated_doc <- add_paragraphs_to_doc(my_doc, text_to_insert) # Insert paragraphs and get the updated document
print(updated_doc, target = "output.docx") # Save Word document
}
}
