% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_to_word.R
\name{add_plot_to_word}
\alias{add_plot_to_word}
\title{Add a plot to a Word document}
\usage{
add_plot_to_word(
  word_doc,
  plot,
  base_width,
  x,
  y,
  orientation = "portrait",
  reference_text = ""
)
}
\arguments{
\item{word_doc}{A Word document object created by the officer package's read_docx() function.}

\item{plot}{A ggplot2 plot object to be added to the Word document.}

\item{base_width}{A numeric value specifying the base width of the plot.}

\item{x}{A numeric value specifying the original width of the plot.}

\item{y}{A numeric value specifying the original height of the plot.}

\item{orientation}{A character string specifying the orientation of the plot. Valid values are "portrait" and "landscape". Default is "portrait".}

\item{reference_text}{A character string containing the reference text to be added after the plot. Default is an empty string.}
}
\value{
A new officer word object that added our plot to the original word object
}
\description{
This function adds a plot to a Word document with a specified base width, aspect ratio, orientation, and reference text.
The function modifies the input Word document directly and does not return any values.
}
\examples{
\dontrun{
# In a Shiny app
output$downloadData <- downloadHandler(
  filename = function() {
    paste("example.docx")
  },
  content = function(file) {
    tmp <- read_docx()

    tmp <- add_plot_to_word(tmp, rocPlot(), 6, x, y, "portrait", "参考文献：Sing, T., Sander, O., Beerenwinkel, N., & Lengauer, T. (2015). ROCR: Visualizing the Performance of Scoring Classifiers. [R package]. Retrieved from https://cran.r-project.org/package=ROCR.")

    tmp \%>\% print(target = file)
  },
  contentType = "application/docx"
)
}

}
