% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots_to_word.R
\name{add_plots_to_word}
\alias{add_plots_to_word}
\title{Add multiple ggplot objects to a Word document}
\usage{
add_plots_to_word(
  word_doc,
  plots,
  base_width,
  x,
  y,
  orientation = "portrait",
  reference_texts = NULL
)
}
\arguments{
\item{word_doc}{An officer Word document object}

\item{plots}{A list of ggplot objects to be added to the Word document}

\item{base_width}{Numeric, the base width of the plots in inches}

\item{x}{Numeric, the width of the plot in pixels}

\item{y}{Numeric, the height of the plot in pixels}

\item{orientation}{Character, the orientation of the page, either "portrait" or "landscape" (default: "portrait")}

\item{reference_texts}{Character vector or NULL, the reference texts to be added after all the plots (default: NULL)}
}
\value{
The modified officer Word document object with the ggplot objects added
}
\description{
This function adds multiple ggplot objects to a Word document with a specified orientation and reference texts.
Reference texts will be displayed after all the plots, with a blank line between them.
}
\examples{

## 下载包含多个图像的Word文档
## 首先，确保已安装并加载所需的包
# install.packages("officer")
# install.packages("flextable")
library(officer)
library(flextable)

## 使用downloadHandler生成下载链接
output$downloadData <- downloadHandler(
  filename = function() {
    paste("example.docx")
  },
  content = function(file) {
    tmp <- read_docx()

    # 创建一个包含多个图像的列表
    plots_list <- list(
      rocPlot(),
      rocPlot(),
      rocPlot()
    )

    reference_text <- "参考文献：Sing, T., Sander, O., Beerenwinkel, N., & Lengauer, T. (2015). ROCR: Visualizing the Performance of Scoring Classifiers. [R package]. Retrieved from https://cran.r-project.org/package=ROCR."

    # 使用add_plots_to_word()函数添加多个图像
    tmp <- add_plots_to_word(tmp, plots = plots_list, base_width = 6, x = x, y = y, orientation = "portrait", reference_texts = reference_text)

    tmp \%>\% print(target = file)
  },
  contentType = "application/docx"
)
}
