% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_to_word.R
\name{add_table_to_word}
\alias{add_table_to_word}
\title{Add a table to a Word document}
\usage{
add_table_to_word(
  word_doc,
  table,
  orientation = "portrait",
  reference_text = ""
)
}
\arguments{
\item{word_doc}{A Word document object created by the officer package's read_docx() function.}

\item{table}{A flextable object to be added to the Word document.}

\item{orientation}{A character string specifying the orientation of the table. Valid values are "portrait" and "landscape". Default is "portrait".}

\item{reference_text}{A character string containing the reference text to be added after the table. Default is an empty string.}
}
\description{
This function adds a table to a Word document with a specified orientation and reference text.
The function modifies the input Word document directly and does not return any values.
}
\examples{
\dontrun{
# In a Shiny app
output$downloadData <- downloadHandler(
  filename = function() {
    paste("example.docx")
  },
  content = function(file) {
    tmp <- read_docx()

    add_table_to_word(tmp, durbin_table(), "portrait", "参考文献：Fox, J., & Weisberg, S. (2020). car: Companion to Applied Regression. [R package]. Retrieved from https://cran.r-project.org/package=car.")

    add_table_to_word(tmp, collin_table(), "portrait", "参考文献：Fox, J., & Weisberg, S. (2020). car: Companion to Applied Regression. [R package]. Retrieved from https://cran.r-project.org/package=car.")

    tmp \%>\% print(target = file)
  },
  contentType = "application/docx"
)
}

}
