% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables_to_word.R
\name{add_tables_to_word}
\alias{add_tables_to_word}
\title{Add multiple tables to a Word document}
\usage{
add_tables_to_word(
  word_doc,
  tables_list,
  orientation = "portrait",
  reference_text = ""
)
}
\arguments{
\item{word_doc}{A Word document object created using officer::read_docx().}

\item{tables_list}{A list of flextable objects to be added to the Word document.}

\item{orientation}{Character, either "portrait" or "landscape". Default is "portrait".}

\item{reference_text}{A character string containing the reference text to be added after each table.}
}
\value{
A modified Word document object with the tables and reference text added.
}
\description{
This function adds multiple tables to a Word document. Each table is followed by the same reference text.
The tables are provided as a list of flextable objects. The function also allows you to set the orientation
of the Word document (either portrait or landscape).
}
\examples{
\dontrun{
library(officer)
library(flextable)
library(magrittr)

# Create a sample Word document
my_doc <- read_docx()

# Create some sample flextables
table1 <- flextable(head(mtcars))
table2 <- flextable(head(iris))
table_list <- list(table1, table2)

# Add the tables to the Word document
my_doc <- add_tables_to_word(my_doc, tables_list = table_list,
                             orientation = "portrait",
                             reference_text = "Sample reference text.")

# Save the Word document
my_doc \%>\% print(target = "example.docx")
}
}
