% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_downloadPlot.R
\name{adjusted_downloadPlot_server}
\alias{adjusted_downloadPlot_server}
\title{Server logic for downloading a plot in various formats}
\usage{
adjusted_downloadPlot_server(id, plot_png, plot_ppt, plot_pdf, x, y)
}
\arguments{
\item{id}{A unique ID for the Shiny module, matching the one used in adjusted_downloadPlot_UI()}

\item{plot_png}{The ggplot2 plot to be downloaded as a PNG}

\item{plot_ppt}{The ggplot2 plot to be downloaded as a PPT}

\item{plot_pdf}{The ggplot2 plot to be downloaded as a PDF}

\item{x}{The width of the plot in pixels}

\item{y}{The height of the plot in pixels}
}
\value{
A module server function
}
\description{
This function handles the server-side logic for downloading a plot in PNG, PPT, and PDF formats.
}
\examples{
# In a Shiny app
# UI part:
adjusted_downloadPlot_UI("plot_downloads")

# Server part:
adjusted_downloadPlot_server("plot_downloads", my_plot_png, my_plot_ppt, my_plot_pdf, 800, 600)

}
