% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_downloadPlot.R
\name{adjusted_downloadPlot_UI}
\alias{adjusted_downloadPlot_UI}
\title{Create UI components for downloading a plot in various formats}
\usage{
adjusted_downloadPlot_UI(id)
}
\arguments{
\item{id}{A unique ID for the Shiny module}
}
\value{
A tagList with the download buttons for the plot
}
\description{
This function creates UI components with download buttons for PNG, PPT, and PDF formats.
}
\examples{
# In a Shiny app
# UI part:
adjusted_downloadPlot_UI("plot_downloads")

# Server part:
adjusted_downloadPlot_server("plot_downloads", my_plot_png, my_plot_ppt, my_plot_pdf, 800, 600)

}
