% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{bernDist}
\alias{bernDist}
\title{Get the probabilities of each possible result in a set of multinomial trials}
\usage{
bernDist(experiments = list())
}
\arguments{
\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The function also accepts a matrix (one row per experiment) In a
\emph{pure} Bernouilli trial, this would be \code{c(number_of_successes,
  number_of_failures)}. The number and order of categories must be the same
in each trial.}
}
\value{
List containing each possible result with the same number of
experiments, tries and outcomes as the input. Each element in the list
contains a matricial description of the result (\code{desc}), the total
number of times each outcome was measured (\code{sdesc}) and the absolute
probability of the result (\code{val}).
}
\description{
This function is called by \link{bernPval}, and therefore is not usually
called independently. It is exported so that interested users can check the
distribution. To also compute the cumulative probabilities, the result must
be sent to \link{calCumul}. A given result can be retrieved with
\link{probeDist}
}
\examples{
r <- bernDist(list(c(8, 2), c(4, 7)))
}
