% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{bernPval}
\alias{bernPval}
\title{p-val calculation to compare mulinomial trials. This function is R only,
use \code{m.test} for a faster alternative.}
\usage{
bernPval(experiments = list())
}
\arguments{
\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The number and order of categories must be the same
in every experiment.
In a
\emph{binomial} experiment, this would be \code{c(number_of_successes,
  number_of_failures)}. The function also accepts a matrix (one column per
experiment).}
}
\value{
p-value according to the null hypothesis that the probability of each
outcome is the same in every experiment (two-sided)
}
\description{
p-val calculation to compare mulinomial trials. This function is R only,
use \code{m.test} for a faster alternative.
}
\examples{
bernPval(list(c(8, 2), c(4, 7)))
}
