% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{calCumul}
\alias{calCumul}
\title{Calculate the cumulative probability of each result}
\usage{
calCumul(l)
}
\arguments{
\item{l}{List containing each possible result with the same number of
experiments, tries and outcomes as the original input. Usually, this will
be the output of \link{bernDist}.}
}
\value{
List with the same elements and fields as the input, plus a
\code{cval} field with the cumulative probability.
}
\description{
The cumulative probability is the sum of the probability of a given result
plus the probabilities of every other equally or less likely result. For a
given result, this is the \code{p-val} returned by \link{bernPval}.
}
\examples{
r <- bernDist(list(c(8, 2), c(4, 7)))
s <- calCumul(r)
}
