% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDT.R
\name{downloadDT}
\alias{downloadDT}
\title{Render a DT table with download buttons}
\usage{
downloadDT(dt, title)
}
\arguments{
\item{dt}{A data frame or tibble to be rendered as a DT table}

\item{title}{A string representing the title of the table, which will be displayed as the table caption}
}
\value{
A DT table with download buttons
}
\description{
This function renders a DT table with buttons to download the table in various formats, such as CSV and Excel.
It also provides options to download the current page or all data.
}
\examples{
# In a Shiny app
# UI part:
DTOutput("orig_data")

# Server part:
output$orig_data <- DT::renderDataTable({
  req(orig_data()) # Ensure the dataset is loaded
  mtest::downloadDT(orig_data(), "The original data is as follows")}, server = FALSE)

}
