% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_numeric_factors.R
\name{extract_numeric_factors}
\alias{extract_numeric_factors}
\title{Extract Numeric Factor Variables from a Dataset}
\usage{
extract_numeric_factors(data)
}
\arguments{
\item{data}{A data frame containing the dataset to be processed.}
}
\value{
A character vector containing the names of factor variables with purely numeric levels.
}
\description{
This function takes a dataset as input and returns a vector containing the names of all factor variables with purely numeric levels.
}
\examples{
data(iris)
numeric_factors <- extract_numeric_factors(iris)
print(numeric_factors)
}
