% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_data.R
\name{fillServer}
\alias{fillServer}
\title{Perform server-side operations for data imputation.}
\usage{
fillServer(input, output, session, recoded)
}
\arguments{
\item{input}{A list of input values from the Shiny UI.}

\item{output}{A list of output values to be displayed on the Shiny UI.}

\item{session}{The Shiny session object.}

\item{recoded}{The dataset to be imputed.}
}
\value{
A reactive expression representing the imputed dataset.
}
\description{
This function performs server-side operations for the data imputation task in Shiny.
It processes the user inputs from the UI, performs data imputation, and renders
the output for the user.
}
\examples{
# This function is designed to be used in a Shiny app and thus cannot be run directly.
# Below is an example of how to use it in a Shiny app.
#
# shiny::shinyServer(function(input, output, session) {
#   imported_data <- mtest::uploadServer(input, output, session, "uploadId")
#   modified <- mtest::modifyServer(input, output, session, "vars", imported_data)
#   recoded <- mtest::recodeServer(input, output, session, "recodeId", modified)
#   imputed_data <- mtest::fillServer(input, output, session, recoded)
# })
}
