% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_data.R
\name{fillUI}
\alias{fillUI}
\title{Generate UI for data imputation.}
\usage{
fillUI()
}
\value{
A shiny fluidPage object that represents the UI of the data imputation task.
}
\description{
This function generates a Shiny UI for the data imputation task. It allows
the user to input various parameters for imputation and view the imputed data.
}
\examples{
# This function is designed to be used in a Shiny app and thus cannot be run directly.
# Below is an example of how to use it in a Shiny app.
#
# ui <- shiny::fluidPage(
#   mtest::fillUI()
# )
}
