% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filterServer}
\alias{filterServer}
\title{Server-side function for data filtering}
\usage{
filterServer(input, output, session, imputed_data)
}
\arguments{
\item{input}{Input parameters from Shiny}

\item{output}{Output parameters to Shiny}

\item{session}{Shiny session information}

\item{imputed_data}{A reactive expression for the data that should be filtered}
}
\value{
data_filtered A reactive expression for the filtered data
}
\description{
Server-side function for data filtering
}
\examples{
# Example usage within a Shiny server function
server <- function(input, output, session) {
  # The following are placeholders and should be replaced with your own data/variables
  imported_data <- mtest::uploadServer(input, output, session, "uploadId")
  modified <- mtest::modifyServer(input, output, session, "vars", imported_data)
  recoded <- mtest::recodeServer(input, output, session, "recodeId", modified)
  imputed_data <- mtest::fillServer(input, output, session, recoded)
  # Use the filterServer function with the imputed data
  filtered_data <- mtest::filterServer (input, output, session, imputed_data)
}
}
