% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filterUI}
\alias{filterUI}
\title{User interface for data filtering}
\usage{
filterUI()
}
\value{
A fluidPage object that provides an interactive data filtering interface
}
\description{
User interface for data filtering
}
\examples{
# Run the filterUI function to generate a data filtering interface
ui_missing <- filterUI()
}
