% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{impute_rules}
\alias{impute_rules}
\title{Impute values based on rules}
\usage{
impute_rules(
  data,
  cat_vars,
  num_vars,
  cat_fill,
  num_fill,
  fixed_value_num,
  fixed_value_cat,
  inplace,
  suffix
)
}
\arguments{
\item{data}{A data frame}

\item{cat_vars}{A vector of categorical variables in the data}

\item{num_vars}{A vector of numerical variables in the data}

\item{cat_fill}{A character string specifying the fill method for categorical variables}

\item{num_fill}{A character string specifying the fill method for numerical variables}

\item{fixed_value_num}{A numeric value used to fill NA values in numerical variables}

\item{fixed_value_cat}{A character value used to fill NA values in categorical variables}

\item{inplace}{A logical value indicating whether to replace the original variables in the data}

\item{suffix}{A character string to append to the new variables}
}
\value{
data The data frame with imputed values
}
\description{
Impute values based on rules
}
\examples{
# Prepare data
data <- data.frame(cat = factor(c("A", "B", "A", NA)), num = c(1, 2, 3, NA))
# Impute values
impute_rules(data, "cat", "num", "纵向缺失值用上一个值替换", "纵向缺失值用上一个值替换", 999, "Unknown", TRUE, "_imputed")
}
