% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{impute_statistical}
\alias{impute_statistical}
\title{Impute values based on statistical measures}
\usage{
impute_statistical(
  data,
  cat_vars,
  num_vars,
  cat_fill,
  num_fill,
  inplace,
  suffix
)
}
\arguments{
\item{data}{A data frame}

\item{cat_vars}{A vector of categorical variables in the data}

\item{num_vars}{A vector of numerical variables in the data}

\item{cat_fill}{A character string specifying the fill method for categorical variables}

\item{num_fill}{A character string specifying the fill method for numerical variables}

\item{inplace}{A logical value indicating whether to replace the original variables in the data}

\item{suffix}{A character string to append to the new variables}
}
\value{
data The data frame with imputed values
}
\description{
Impute values based on statistical measures
}
\examples{
# Prepare data
data <- data.frame(cat = factor(c("A", "B", "A", NA)), num = c(1, 2, 3, NA))
# Impute values
impute_statistical(data, "cat", "num", "众数填充", "均值填充", TRUE, "_imputed")
}
