% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_variable_selection.R
\name{lasso_variable_selection}
\alias{lasso_variable_selection}
\title{Lasso variable selection}
\usage{
lasso_variable_selection(
  df,
  response_var,
  predictor_vars,
  family = "gaussian",
  lambda_selection = "lambda.1se",
  alpha = 1,
  nlambda = 100,
  nfolds = 10
)
}
\arguments{
\item{df}{A data frame.}

\item{response_var}{A string, the name of the response variable in the data frame.}

\item{predictor_vars}{A vector of strings, the names of the predictor variables in the data frame.}

\item{family}{A string, the model family to be used in the lasso regression. Default is "gaussian".}

\item{lambda_selection}{A string, the criterion to select the best lambda value. Can be either "lambda.1se" or "lambda.min". Default is "lambda.1se".}

\item{alpha}{A numeric value, the alpha parameter for the lasso regression. Default is 1.}

\item{nlambda}{A numeric value, the number of lambda values to be used in the lasso regression. Default is 100.}

\item{nfolds}{A numeric value, the number of folds to be used in the cross-validation. Default is 10.}
}
\value{
A list with the following components:
\itemize{
 \item "coef_table": A data frame with the coefficient table from the lasso regression.
 \item "non_zero_vars": A vector of strings, the names of the predictor variables with non-zero coefficients.
 \item "cv_plot": A plot object, the cross-validation plot from the lasso regression.
 \item "lasso_plot": A plot object, the lasso path plot from the lasso regression.
}
}
\description{
This function uses lasso regression to perform variable selection.
}
\examples{
# Use mtcars dataset as an example
df <- mtcars
response_var <- "mpg"
predictor_vars <- setdiff(names(df), response_var)

# Run lasso variable selection
result <- lasso_variable_selection(df, response_var, predictor_vars)

# Print results
print(result$coef_table)
print(result$non_zero_vars)

# Show plots
replayPlot(result$cv_plot)
replayPlot(result$lasso_plot)

}
