% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lest2.R
\name{Lest2}
\alias{Lest2}
\title{Modified Ripley's L(r)}
\usage{
Lest2(...)
}
\arguments{
\item{...}{Passes arguments to Kest() from spatstat}
}
\value{
A vector of L and r
}
\description{
Placeholder documentation: A Modification of Ripley's L: sqrt(K/pi)-R
}
