% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{m.test}
\alias{m.test}
\title{p-val calculation to compare multinomial tests}
\usage{
m.test(experiments, null.hypothesis = "equal")
}
\arguments{
\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The function also accepts a matrix (one row per experiment) In a
\emph{pure} Bernouilli trial, this would be \code{c(number_of_successes,
  number_of_failures)}. The number and order of categories must be the same
in each trial.}

\item{null.hypothesis}{a character string specifying the alternative hypothesis,
must be one of "equal" (default), "higher" or "lower".}
}
\value{
p-value for the null hypothesis (all underlying probabilities
are the same in every experiment)
}
\description{
p-val calculation to compare multinomial tests
}
\examples{
m.test(list(c(8, 2), c(4, 7)))
}
