% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_test.R
\name{M_test}
\alias{M_test}
\title{Monte Carloe test of differences between medians}
\usage{
M_test(x, group, nperm = 999, ...)
}
\arguments{
\item{x}{quantitative values to be tested}

\item{group}{a factor defining groups; same length as 'x'}

\item{nperm}{the number of permutations to be used to generate the reference
distribution of the statistic}
}
\description{
The 'M test' tests the median differences between two groups. This test is
non-parametric, and uses a Monte-Carlo procedure to approximate the null
distribution of the statistic.
}
\examples{

head(VL)
test <- M_test(VL$titer, VL$type, alter = "less")
test
plot(test)

}
