% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{mcm}
\alias{mcm}
\title{Monte Carlo simulation of multinomial process}
\usage{
mcm(experiments = list(), n = 10000, algo = 1, NRUNIF = 0)
}
\arguments{
\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The function also accepts a matrix (one row per experiment) In a
\emph{pure} Bernouilli trial, this would be \code{c(number_of_successes,
  number_of_failures)}. The number and order of categories must be the same
in each trial.}

\item{n}{Number of simulations to run}

\item{algo}{Algorithm for pseudorandom matrix}
}
\value{
List of results, as described in \link{bernDist}
}
\description{
Monte Carlo simulation of multinomial process
}
\examples{
mcm(list(c(4, 5), c(6, 7)), n=1000)
}
