% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mest.R
\name{Mest}
\alias{Mest}
\title{M(r) function}
\usage{
Mest(X1, X2, r = NULL, rmax = NULL)
}
\arguments{
\item{X1}{A point pattern object (ppp)}

\item{X2}{Another point pattern object (ppp)}

\item{r}{vector of distances at which to evaluate M(r)}

\item{rmax}{the maximum distance at which to evaluate M(r)}
}
\value{
returns a vector of M values for a series of distances (r)
}
\description{
Calculates the difference of the Ripley's L(r) of two point patterns
}
