% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_data.R
\name{modifyServer}
\alias{modifyServer}
\title{Modify Variables Server}
\usage{
modifyServer(input, output, session, id, importData)
}
\arguments{
\item{input, }{output, session The input, output, and session objects of the Shiny server.}

\item{id}{A string, the namespace of the module.}

\item{importData}{A dataframe, the data to be modified.}
}
\value{
Returns a reactive expression that expresses the updated data.
}
\description{
This function generates a Shiny server module that allows the user to modify variables in a dataset.
}
\details{
修改变量的Server端
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    mtest::modifyUI("vars")
  )
  server <- function(input, output, session) {
    imported_data <- mtest::uploadServer(input, output, session, "uploadId")
    modified <- mtest::modifyServer(input, output, session, "vars", imported_data)
  }
  shinyApp(ui, server)
}
}
