% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplot.R
\name{Mplot}
\alias{Mplot}
\title{Plotting function for M-test}
\usage{
Mplot(mtest, ...)
}
\arguments{
\item{mtest}{the output of m.test()}

\item{...}{pass arguments to plot()}
}
\value{
Plot the observed M(r) versus the randomization envelope
}
\description{
Makes a pretty plot
}
