% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nongg_downloadPlot.R
\name{nongg_downloadPlot_server}
\alias{nongg_downloadPlot_server}
\title{Server logic for downloading a plot in various formats}
\usage{
nongg_downloadPlot_server(
  id,
  plot_png,
  plot_ppt,
  plot_pdf,
  x,
  y,
  render = "plot"
)
}
\arguments{
\item{id}{A unique ID for the Shiny module, matching the one used in adjusted_downloadPlot_UI()}

\item{plot_png}{The ggplot2 plot to be downloaded as a PNG}

\item{plot_ppt}{The ggplot2 plot to be downloaded as a PPT}

\item{plot_pdf}{The ggplot2 plot to be downloaded as a PDF}

\item{x}{The width of the plot in pixels}

\item{y}{The height of the plot in pixels}

\item{render}{A string, either "plot" or "print". When "plot", it uses plot() function to render the plot. When "print", it uses print() function to render the plot.}
}
\value{
A module server function
}
\description{
This function handles the server-side logic for downloading a plot in PNG, PPT, and PDF formats.
}
\examples{
# In a Shiny app
# UI part:
nongg_downloadPlot_UI("plot_downloads")

# Server part:
nongg_downloadPlot_server("plot_downloads", my_plot_png, my_plot_ppt, my_plot_pdf, 800, 600, "print")

}
