% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_data.R
\name{recodeServer}
\alias{recodeServer}
\title{Recode Variables Server}
\usage{
recodeServer(input, output, session, id, data)
}
\arguments{
\item{input, }{output, session The input, output, and session objects of the Shiny server.}

\item{id}{A string, the namespace of the module.}

\item{data}{A reactive, the data to be recoded.}
}
\value{
Returns a reactive expression that expresses the recoded data.
}
\description{
This function generates a Shiny server module for recoding categorical variables.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    mtest::recodeUI("recodeId")
  )
  server <- function(input, output, session) {
    data <- reactive({ data.frame() })
    recoded <- mtest::recodeServer(input, output, session, "recodeId", data)
  }
  shinyApp(ui, server)
}
}
