% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_data.R
\name{recodeUI}
\alias{recodeUI}
\title{Recode Variables UI}
\usage{
recodeUI(id)
}
\arguments{
\item{id}{A string, the namespace of the module.}
}
\value{
Returns a shiny UI object that can be used in UI assembly.
}
\description{
This function generates a Shiny UI module for recoding categorical variables.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    mtest::recodeUI("recodeId")
  )
  server <- function(input, output, session) { }
  shinyApp(ui, server)
}
}
