% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_data.R
\name{relevelServer}
\alias{relevelServer}
\title{Server-side Function for Factor Level Reordering}
\usage{
relevelServer(input, output, session, data)
}
\arguments{
\item{input}{Input parameters from Shiny}

\item{output}{Output parameters to Shiny}

\item{session}{Shiny session information}

\item{data}{A reactive expression for the data that should be reordered}
}
\value{
ordered A reactive expression for the reordered data
}
\description{
Server-side Function for Factor Level Reordering
}
\examples{
# Example usage within a Shiny server function
server <- function(input, output, session) {
  # The following are placeholders and should be replaced with your own data/variables
  imported_data <- mtest::uploadServer(input, output, session, "uploadId")
  modified <- mtest::modifyServer(input, output, session, "vars", imported_data)
  recoded <- mtest::recodeServer(input, output, session, "recodeId", modified)
  imputed_data <- mtest::fillServer(input, output, session, recoded)
  filtered_data <- mtest::filterServer (input, output, session, imputed_data)
  # Use the relevelServer function with the filtered data
  ordered <- mtest::relevelServer (input, output, session, filtered_data)
}
}
