% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_data.R
\name{relevelUI}
\alias{relevelUI}
\title{User Interface for Factor Level Reordering}
\usage{
relevelUI()
}
\value{
A fluidPage object that provides an interactive interface for reordering factor levels
}
\description{
User Interface for Factor Level Reordering
}
\examples{
# Run the relevelUI function to generate an interface for reordering factor levels
ui_factor <- mtest::relevelUI()
}
