% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots_UI.R
\name{render_plots_UI}
\alias{render_plots_UI}
\title{Render a list of plots in a Shiny UI}
\usage{
render_plots_UI(plots, title_fun)
}
\arguments{
\item{plots}{A list of plot objects to be rendered}

\item{title_fun}{A function that takes an integer i and returns a string for the title of the i-th plot}
}
\value{
A tagList object containing the rendered plots and titles
}
\description{
This function renders a list of plots in a Shiny UI, along with their titles.
}
\examples{
\dontrun{
# In a Shiny app
output$plots <- renderUI({
  req(input$OK)
  isolate({
    req(length(input$comp) > 0 || length(input$solo) > 0)
    render_plots_UI(plots(), function(i) {
      glue("Estimated Marginal Means for {input$dep} by {input$emMeans[[i]]}")
    })
  })
})
}
}
