\name{simDataMTest}
\docType{data}
\alias{simDataMTest}
\title{Simulated data for MTest}
\description{
  This data set helps testing functions in MTest package, the generating process is documented in the reference.
}
\usage{simDataMTest}
\format{A dataframe containing 10000 observations and four columns.}
\references{
  Morales-Oñate, V., and Morales-Oñate, B. (2023). \emph{MTest: a Bootstrap Test for Multicollinearity}. Revista Politécnica, 51(2), 53–62. \doi{https://doi.org/10.33333/rp.vol51n2.05}
}
\keyword{datasets}