% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_variables.R
\name{split_variables}
\alias{split_variables}
\title{Split Variables Function}
\usage{
split_variables(data, variables)
}
\arguments{
\item{data}{A data frame in which to look for the variables.}

\item{variables}{A character vector of variable names.}
}
\value{
A list containing two character vectors of variable names: `covs` for numeric variables and `factors` for non-numeric variables.
}
\description{
This function splits the provided variable names into two categories: numeric and non-numeric.
}
\examples{
\dontrun{
data <- data.frame(A = rnorm(100), B = letters[1:100], C = rnorm(100), D = letters[1:100])
vars <- c("A", "B", "C", "D")
result <- split_variables(data, vars)
numeric_vars <- result$covs
non_numeric_vars <- result$factors
print(numeric_vars)
print(non_numeric_vars)
}
}
