% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{uploadUI}
\alias{uploadUI}
\title{Upload UI for Shiny}
\usage{
uploadUI(id, example_data)
}
\arguments{
\item{id}{A character string for the id of the UI elements}
}
\value{
A shiny UI definition for file upload
}
\description{
A function to generate a Shiny UI for uploading files
}
\examples{
if (interactive()) {
  shinyApp(
    ui = fluidPage(
      uploadUI("uploadId","example.csv")
    ),
    server = function(input, output, session) {
      imported <- uploadServer("uploadId")
    }
  )
}
}
