% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mtest.R
\name{m.test}
\alias{m.test}
\title{M-test}
\usage{
m.test(X1, X2, b = 100)
}
\arguments{
\item{X1}{A point pattern object (ppp)}

\item{X2}{Another point pattern object (ppp)}

\item{b}{number of bootstrap replicates to generate randomization envelope}
}
\value{
Returns a list of things
}
\description{
Does that thang
}
