% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{m.test}
\alias{m.test}
\title{p-val calculation to compare multinomial trials}
\usage{
m.test(experiments)
}
\arguments{
\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The number and order of categories must be the same
in every experiment.
In a
\emph{binomial} experiment, this would be \code{c(number_of_successes,
  number_of_failures)}. The function also accepts a matrix (one column per
experiment).}
}
\value{
p-value for the null hypothesis (all underlying outcome probabilities
are the same in every experiment)
}
\description{
p-val calculation to compare multinomial trials
}
\examples{
m.test(list(c(8, 2), c(4, 7)))
}
