% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{probeDist}
\alias{probeDist}
\title{Check a given result in a distribution}
\usage{
probeDist(l, experiments)
}
\arguments{
\item{l}{List containing each possible result with the same number of
experiments, tries and outcomes as the original input. Usually, this will
be the output of \link{bernDist}.}

\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The function also accepts a matrix (one row per experiment) In a
\emph{pure} Bernouilli trial, this would be \code{c(number_of_successes,
  number_of_failures)}. The number and order of categories must be the same
in each trial.}
}
\value{
Data calculated for \code{experiment} in \code{l}. This includes the
matricial description of the result (\code{desc}), the total number of
times each outcome was measured (\code{sdesc}) and the absolute probability
of the result (\code{val}). If \link{calCumul} has been run, it will also
show the cumulative probability (\code{cval}). If the \code{experiment} is
not found in \code{l}, the function returns \code{NULL}.
}
\description{
Once a distribution of results has been calculated with \link{bernDist} and
possibly \link{calCumul}, this function returns information on a particular
result
}
\examples{
r <- bernDist(list(c(8, 2), c(4, 7)))
s <- calCumul(r)
t <- probeDist(s, list(c(8, 2), c(4, 7)))
}
