% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots_UI.R
\name{render_plots_UI}
\alias{render_plots_UI}
\title{Add a table to a Word document}
\usage{
render_plots_UI(plots, title_glue)
}
\arguments{
\item{plots}{A list of ggplot objects to be rendered in the Shiny app.}

\item{titles}{A character string containing the title text to be displayed above each plot. Can include placeholders for dynamic values.}
}
\value{
A Shiny UI element containing the rendered plots and their titles.
}
\description{
This function adds a table to a Word document with a specified orientation and reference text.
The function modifies the input Word document directly and does not return any values.
}
\examples{
\dontrun{
# In a Shiny app
output$plots <- renderUI({
  req(input$OK)
  isolate({
    req(length(input$comp) > 0 || length(input$solo) > 0)
    render_plots_UI(plots(), "Estimated Marginal Means for {input$dep} by {input$emMeans[[i]]}")
  })
})
}

}
