% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtest.R
\name{tailed.m.test}
\alias{tailed.m.test}
\title{Two-tail p-val calculation to compare binomial tests}
\usage{
tailed.m.test(experiments)
}
\arguments{
\item{experiments}{List of multinomial trials. Each
trial is codified as a vector with the number of results in each category.
The function also accepts a matrix (one row per experiment) In a
\emph{pure} Bernouilli trial, this would be \code{c(number_of_successes,
  number_of_failures)}. The number and order of categories must be the same
in each trial.}
}
\value{
p-value for the null hypothesis (the probability of success in the
second experiment is lower than the probability of success in the first
experiment)
}
\description{
Two-tail p-val calculation to compare binomial tests
}
\examples{
tailed.m.test(list(c(10, 8), c(10, 3)))
}
